+++ dst/app/routes/transactions/placeOrder.js
@@ -22,6 +22,7 @@ const moment = require("moment");
 // tslint:disable-next-line:no-require-imports no-var-requires
 // const httpsAgent = require('agentkeepalive').HttpsAgent;
 // const agent = require('agentkeepalive');
+const WAITER_DISABLED = process.env.WAITER_DISABLED === '1';
 const placeOrderTransactionsRouter = express_1.Router();
 const authentication_1 = require("../../middlewares/authentication");
 const permitScopes_1 = require("../../middlewares/permitScopes");
@@ -47,8 +48,9 @@ placeOrderTransactionsRouter.post('/start', permitScopes_1.default(['transaction
         .withMessage('seller_identifier is required');
     // POSからの流入制限を一時的に回避するため、許可証不要なクライアント設定ができるようにする
     // staffアプリケーションに関しても同様に
-    if (req.user.client_id !== process.env.POS_CLIENT_ID &&
-        req.user.client_id !== process.env.STAFF_CLIENT_ID) {
+    if (!WAITER_DISABLED
+        && req.user.client_id !== process.env.POS_CLIENT_ID
+        && req.user.client_id !== process.env.STAFF_CLIENT_ID) {
         req.checkBody('passportToken', 'invalid passportToken')
             .notEmpty()
             .withMessage('passportToken is required');