+++ dst/app/routes/transactions/returnOrder.js
@@ -47,10 +47,15 @@ returnOrderTransactionsRouter.post('/confirm', permitScopes_1.default(['transact
         const transactionRepo = new ttts.repository.Transaction(ttts.mongoose.connection);
         // 取引を検索する
         const conditions = {
-            result: { $exists: true },
             typeOf: ttts.factory.transactionType.PlaceOrder,
-            'result.eventReservations.performance_day': req.body.performance_day,
-            'result.eventReservations.payment_no': req.body.payment_no
+            'result.eventReservations.performance_day': {
+                $exists: true,
+                $eq: req.body.performance_day
+            },
+            'result.eventReservations.payment_no': {
+                $exists: true,
+                $eq: req.body.payment_no
+            }
         };
         debug('searching a transaction...', conditions);
         const placeOrderTransaction = yield transactionRepo.transactionModel.findOne(conditions)