+++ dst/app/app.js
@@ -15,6 +15,7 @@ const admins_1 = require("./routes/admins");
 const dev_1 = require("./routes/dev");
 const events_1 = require("./routes/events");
 const health_1 = require("./routes/health");
+const iam_1 = require("./routes/iam");
 const oauth_1 = require("./routes/oauth");
 const orders_1 = require("./routes/orders");
 const organizations_1 = require("./routes/organizations");
@@ -68,6 +69,7 @@ app.use(expressValidator({})); // this line must be immediately after any of the
 app.use('/admins', admins_1.default);
 app.use('/events', events_1.default);
 app.use('/health', health_1.default);
+app.use('/iam', iam_1.default);
 app.use('/oauth', oauth_1.default);
 app.use('/orders', orders_1.default);
 app.use('/organizations', organizations_1.default);