+++ src/app/routes/transactions/placeOrder.ts
@@ -13,6 +13,8 @@ import * as moment from 'moment';
 // const httpsAgent = require('agentkeepalive').HttpsAgent;
 // const agent = require('agentkeepalive');
 
+const WAITER_DISABLED = process.env.WAITER_DISABLED === '1';
+
 const placeOrderTransactionsRouter = Router();
 
 import authentication from '../../middlewares/authentication';
@@ -50,8 +52,9 @@ placeOrderTransactionsRouter.post(
 
         // POSからの流入制限を一時的に回避するため、許可証不要なクライアント設定ができるようにする
         // staffアプリケーションに関しても同様に
-        if (req.user.client_id !== <string>process.env.POS_CLIENT_ID &&
-            req.user.client_id !== <string>process.env.STAFF_CLIENT_ID) {
+        if (!WAITER_DISABLED
+            && req.user.client_id !== <string>process.env.POS_CLIENT_ID
+            && req.user.client_id !== <string>process.env.STAFF_CLIENT_ID) {
             req.checkBody('passportToken', 'invalid passportToken')
                 .notEmpty()
                 .withMessage('passportToken is required');
@@ -78,7 +81,7 @@ placeOrderTransactionsRouter.post(
                 passportToken: req.body.passportToken
             })(
                 new ttts.repository.Transaction(ttts.mongoose.connection),
-                new ttts.repository.Organization(ttts.mongoose.connection)
+                new ttts.repository.Seller(ttts.mongoose.connection)
             );
 
             // tslint:disable-next-line:no-string-literal
@@ -246,7 +249,7 @@ placeOrderTransactionsRouter.post(
                 creditCard
             )(
                 new ttts.repository.action.authorize.CreditCard(ttts.mongoose.connection),
-                new ttts.repository.Organization(ttts.mongoose.connection),
+                new ttts.repository.Seller(ttts.mongoose.connection),
                 new ttts.repository.Transaction(ttts.mongoose.connection),
                 creditService
             );