+++ src/app/routes/transactions/returnOrder.ts
@@ -49,10 +49,15 @@ returnOrderTransactionsRouter.post(
 
             // 取引を検索する
             const conditions = {
-                result: { $exists: true },
                 typeOf: ttts.factory.transactionType.PlaceOrder,
-                'result.eventReservations.performance_day': req.body.performance_day,
-                'result.eventReservations.payment_no': req.body.payment_no
+                'result.eventReservations.performance_day': {
+                    $exists: true,
+                    $eq: req.body.performance_day
+                },
+                'result.eventReservations.payment_no': {
+                    $exists: true,
+                    $eq: req.body.payment_no
+                }
             };
             debug('searching a transaction...', conditions);
             const placeOrderTransaction = await transactionRepo.transactionModel.findOne(conditions)